move order_channel_by_depth to core
