+++ spikeinterface/core/__init__.py
@@ -58,7 +58,7 @@ from .core_tools import write_binary_recording, write_to_h5_dataset_format, writ
     write_python
 from .job_tools import ensure_n_jobs, ensure_chunk_size, ChunkRecordingExecutor
 from .recording_tools import (get_random_data_chunks, get_channel_distances, get_closest_channels, 
-                              get_noise_levels, get_chunk_with_margin)
+                              get_noise_levels, get_chunk_with_margin, order_channels_by_depth)
 from .waveform_tools import extract_waveforms_to_buffers
 from .snippets_tools import snippets_from_sorting
